+++ lib/chamba_search_mx/factories/base_url.rb
@@ -1,0 +1,5 @@
+module ChambaSearchMx
+  ##
+  # this is a secret!
+  BASE_URL = ENV['CHAMBA_SEARCH_MX_BASE_URL'].freeze
+end