+++ lib/chamba_search_mx/factories/salaries.rb
@@ -1,4 +1,4 @@
-module OccSearch
+module ChambaSearchMx
   ##
   # available salary ranges
   SALARIES = {