+++ lib/chamba_search_mx/html_page.rb
@@ -1,21 +1,21 @@
 require 'nokogiri'
 require 'open-uri'
 
-module OccSearch
+module ChambaSearchMx
   ##
   # loads web page using nokogiri
   class HTMLPage
     attr_reader :content
     ##
     # @param <hash> options
-    # @see OccSearch::BASE_URL
+    # @see ChambaSearchMx::BASE_URL
     def initialize(opts = {})
-      @base_url = opts[:base_url] || OccSearch::BASE_URL
+      @base_url = opts[:base_url] || ChambaSearchMx::BASE_URL
     end
 
     ##
     # loads Kokogiri::HTML:Document as @content which can be read
-    # @param <string> occ url
+    # @param <string>
     def load(url)
       @content = Nokogiri::HTML open("#{@base_url}#{url}")
     end