+++ lib/occ_search/factories/base_url.rb
@@ -1,5 +1,0 @@
-module OccSearch
-  ##
-  # OCC url
-  BASE_URL = 'https://www.occ.com.mx'.freeze
-end