+++ spec/job_page_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-describe OccSearch::JobPage do
+describe ChambaSearchMx::JobPage do
   if ENV['SLOW_TESTS']
     context 'slow tests' do
       search_url = '/empleos-en-jalisco/desde-hace-60-dias/web-developer'
-      search_page = OccSearch::SearchPage.new
+      search_page = ChambaSearchMx::SearchPage.new
       search_page.load search_url
       url = search_page.job_urls.first
-      job_page = OccSearch::JobPage.new
+      job_page = ChambaSearchMx::JobPage.new
       job_page.load(url)
 
       it 'should reponse to load' do
@@ -15,11 +15,11 @@ describe OccSearch::JobPage do
       end
 
       it 'load should return itself' do
-        expect(job_page.load(url)).to be_instance_of(OccSearch::JobPage)
+        expect(job_page.load(url)).to be_instance_of(ChambaSearchMx::JobPage)
       end
 
-      it 'data should be an instance of OccSearch::JobDecorator' do
-        expect(job_page.data).to be_instance_of(OccSearch::JobDecorator)
+      it 'data should be an instance of ChambaSearchMx::JobDecorator' do
+        expect(job_page.data).to be_instance_of(ChambaSearchMx::JobDecorator)
       end
     end
   end