+++ spec/search_page_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe OccSearch::SearchPage do
+describe ChambaSearchMx::SearchPage do
   url = '/empleos-en-jalisco/desde-hace-60-dias/web-developer'
-  page = OccSearch::SearchPage.new
+  page = ChambaSearchMx::SearchPage.new
   page.load url
 
   describe page.job_urls do