+++ spec/spec_helper.rb
@@ -1,3 +1,3 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'byebug'
-require 'occ_search'
+require 'chamba_search_mx'